ported changes to new file
