+++ src/components/Navbar.jsx
@@ -5,12 +5,7 @@ import { navLinks } from '../constants';
 import { logo, menu, close } from '../assets';
 import { fadeIn } from '../utils/motion';
 import { AnimatePresence, motion } from "framer-motion";
-// import Dissertation from "./Dissertation";
-
-const center = {
-  x: window.innerWidth / 2,
-  y: window.innerHeight / 2,
-};
+import Dissertation from "./Dissertation";
 
 const show = {
   opacity: 1,
@@ -23,21 +18,7 @@ const hide = {
       display: "none"
   }
 };
-const Dissertation = ({setter, title}) => {
-  const [isOpen, setIsOpen] = useState(true);
-  return (
-      <motion.div
-        className={`z-20 absolute flex items-center justify-center h-screen left-[5px] right-[5px] top-[80px] bg-indigo-500/10 cursor-default`}
-        initial={{ opacity: 0, scale: 0.5, translateX: -center.x, translateY: -center.y }}
-        animate={isOpen ? { opacity: 1, scale: 1, translateX: 0, translateY: 0 } : { opacity: 0, scale: 1, translateX: 0, translateY: 0 }}
-      >
-           <div className='absolute left-[5px] right-[5px] top-[20px] h-[28px] mr-[16px]'>
-                    <motion.img animate src={close} alt="menu" className='w-[28px] h-[28px] object-contain cursor-pointer absolute' onClick={()=> {setIsOpen(!isOpen); console.log(isOpen);}}/>
-            </div>
-          <p className='backdrop-blur-2xl text-secondary text-[17px] max-w-3xl'>Sed aliquam sollicitudin turpis, quis ullamcorper dolor finibus non. Suspendisse lacus arcu, finibus ac ligula quis, luctus luctus turpis. Quisque eleifend ex et eleifend pretium. Suspendisse ac lacus felis. Sed blandit finibus odio, sed viverra justo lacinia luctus. Aliquam sed arcu non elit malesuada fermentum a quis lorem. Curabitur at dui eleifend, ultricies lorem finibus, elementum nulla. Fusce ipsum mauris, aliquam non odio sed, finibus feugiat purus. Nam ac vehicula turpis, id laoreet massa. Aliquam sed condimentum augue. Ut eu augue pulvinar dolor porta congue. Morbi facilisis at nisl nec cursus. Nulla eu sapien feugiat, viverra augue sit amet, feugiat eros. Pellentesque nec massa enim. Morbi volutpat sollicitudin ante at elementum. Vestibulum pharetra dui at egestas tempor.</p>
-      </motion.div>
-  )
-}
+
 
 const Navbar = () => {
   const [active, setActive] = useState('');
@@ -54,10 +35,10 @@ const Navbar = () => {
         </Link>
         <ul className='list-none hidden sm:flex flex-row gap-10'>
           {navLinks.map((link) => (
-            <li key={link.id} className={`${active === link.title ? "text-white" : "text-secondary"} hover:text-white text-[18px] font-medium cursor-pointer`}
+            <li key={link.id} className={`${(active === link.title && active !== "Dissertation") ? "text-white" : "text-secondary"} hover:text-white text-[18px] font-medium cursor-pointer`}
                 onClick={()=> setActive(link.title)}>
                 <a href={`#${link.id}`}>{link.title}</a>
-                {(active==="Dissertation") && (link.id==="dissertation") ? <Dissertation currentSetter={setActive} currentTitle={link.title}/> : <></>}
+                {(active==="Dissertation") && (link.id==="dissertation") ? <Dissertation/> : <></>}
             </li>
           ))}
         </ul>
